Add stories for SidebarBottom and move into components folder
