+++ src/screens/VisualTests/VisualTests.stories.tsx
@@ -3,15 +3,7 @@
 import { VariablesOf } from "@graphql-typed-document-node/core";
 import type { Meta, StoryObj } from "@storybook/react";
 import { expect, fn } from "@storybook/test";
-import {
-  findByLabelText,
-  findByRole,
-  fireEvent,
-  screen,
-  userEvent,
-  waitFor,
-  within,
-} from "@storybook/testing-library";
+import { findByLabelText, findByRole, fireEvent, waitFor } from "@storybook/testing-library";
 import { delay, HttpResponse } from "msw";
 import React from "react";
